+++ packages/xchain-sui/src/index.ts
@@ -2,3 +2,4 @@ export { Client } from './client'
 
 export * from './types'
 export * from './const'
+export { getDefaultClientUrl, getDefaultGraphqlUrl, getDefaultGrpcUrl, getSuiNetwork } from './utils'